+++ packages/svelte/src/components/Form.svelte
@@ -1,10 +1,12 @@
 <script lang="ts">
   import {
     formDataToObject,
+    FormComponentResetSymbol,
     resetFormFields,
     mergeDataIntoQueryString,
     type Errors,
     type FormComponentProps,
+    type FormComponentRef,
     type Method,
     type FormDataConvertible,
     type VisitOptions,
@@ -13,7 +15,9 @@
   } from '@inertiajs/core'
   import { type NamedInputEvent, type ValidationConfig, type Validator } from 'laravel-precognition'
   import { isEqual } from 'lodash-es'
-  import { onMount } from 'svelte'
+  import { onMount, setContext } from 'svelte'
+  import { writable } from 'svelte/store'
+  import { FormContextKey } from './formContext'
   import useForm from '../useForm'
 
   const noop = () => undefined
@@ -92,6 +96,11 @@
   }
 
   function updateDirtyState(event: Event) {
+    if (event.type === 'reset' && (event as CustomEvent).detail?.[FormComponentResetSymbol]) {
+      // When the form is reset programmatically, prevent native reset behavior
+      event.preventDefault()
+    }
+
     isDirty = event.type === 'reset' ? false : !isEqual(getData(), formDataToObject(defaultData))
   }
 
@@ -223,7 +232,7 @@
 
     form.defaults(getData())
 
-    const formEvents = ['input', 'change', 'reset']
+    const formEvents: Array<keyof HTMLElementEventMap> = ['input', 'change', 'reset']
 
     formEvents.forEach((e) => formElement.addEventListener(e, updateDirtyState))
 
@@ -243,6 +252,37 @@
   }
 
   $: slotErrors = $form.errors as Errors
+
+  // Form context for child components
+  const formContextStore = writable<FormComponentRef | undefined>(undefined)
+
+  $: formContextStore.set({
+    errors: $form.errors,
+    hasErrors: $form.hasErrors,
+    processing: $form.processing,
+    progress: $form.progress,
+    wasSuccessful: $form.wasSuccessful,
+    recentlySuccessful: $form.recentlySuccessful,
+    isDirty,
+    clearErrors,
+    resetAndClearErrors,
+    setError,
+    reset,
+    submit,
+    defaults,
+    getData,
+    getFormData,
+    // Precognition
+    validator,
+    validate,
+    touch,
+    validating: $form.validating,
+    valid,
+    invalid,
+    touched,
+  })
+
+  setContext(FormContextKey, formContextStore)
 </script>
 
 <form