+++ packages/vue3/src/form.ts
@@ -2,6 +2,7 @@ import {
   Errors,
   FormComponentProps,
   FormComponentRef,
+  FormComponentResetSymbol,
   FormComponentSlotProps,
   FormDataConvertible,
   formDataToObject,
@@ -14,13 +15,28 @@ import {
 } from '@inertiajs/core'
 import { NamedInputEvent, ValidationConfig } from 'laravel-precognition'
 import { isEqual } from 'lodash-es'
-import { computed, defineComponent, h, onBeforeUnmount, onMounted, PropType, ref, SlotsType, watch } from 'vue'
+import {
+  computed,
+  defineComponent,
+  h,
+  inject,
+  InjectionKey,
+  onBeforeUnmount,
+  onMounted,
+  PropType,
+  provide,
+  ref,
+  SlotsType,
+  watch,
+} from 'vue'
 import useForm from './useForm'
 
 type FormSubmitOptions = Omit<VisitOptions, 'data' | 'onPrefetched' | 'onPrefetching'>
 type FormSubmitter = HTMLElement | null
 
 const noop = () => undefined
+
+const FormContextKey: InjectionKey<FormComponentRef> = Symbol('InertiaFormContext')
 
 const Form = defineComponent({
   name: 'Form',
@@ -163,9 +179,11 @@ const Form = defineComponent({
     const defaultData = ref(new FormData())
 
     const onFormUpdate = (event: Event) => {
-      // If the form is reset, we set isDirty to false as we already know it's back
-      // to defaults. Also, the fields are updated after the reset event, so the
-      // comparison will be incorrect unless we use nextTick/setTimeout.
+      if (event.type === 'reset' && (event as CustomEvent).detail?.[FormComponentResetSymbol]) {
+        // When the form is reset programmatically, prevent native reset behavior
+        event.preventDefault()
+      }
+
       isDirty.value = event.type === 'reset' ? false : !isEqual(getData(), formDataToObject(defaultData.value))
     }
 
@@ -325,6 +343,8 @@ const Form = defineComponent({
 
     expose<FormComponentRef>(exposed)
 
+    provide(FormContextKey, exposed)
+
     return () => {
       return h(
         'form',
@@ -344,5 +364,9 @@ const Form = defineComponent({
     }
   },
 })
+
+export function useFormContext(): FormComponentRef | undefined {
+  return inject(FormContextKey)
+}
 
 export default Form