+++ packages/vue3/test-app/Pages/FormComponent/Context/NestedComponent.vue
@@ -1,0 +1,7 @@
+<script setup lang="ts">
+import DeeplyNestedComponent from './DeeplyNestedComponent.vue'
+</script>
+
+<template>
+  <DeeplyNestedComponent />
+</template>