+++ packages/vue3/test-app/Pages/Prefetch/AfterError.vue
@@ -8,11 +8,21 @@ const prefetchPage = () => {
 const visitPage = () => {
   router.visit('/prefetch/swr/1')
 }
+
+const prefetchNonInertia = () => {
+  router.prefetch('/non-inertia', { method: 'get' }, { cacheFor: 5000 })
+}
+
+const visitNonInertia = () => {
+  router.visit('/non-inertia')
+}
 </script>
 
 <template>
   <div>
     <button @click="prefetchPage">Prefetch Page</button>
     <button @click="visitPage">Visit Page</button>
+    <button @click="prefetchNonInertia">Prefetch Non-Inertia</button>
+    <button @click="visitNonInertia">Visit Non-Inertia</button>
   </div>
 </template>